refactor: stackingdao data should be in bitcoin-staking-data
