+++ apps/web/app/data/bitcoin-staking-data.ts
@@ -161,3 +161,29 @@ export function isPoolAvailableOnNetwork(
 ): boolean {
   return (pool.signerManagerContracts[networkMode]?.length ?? 0) > 0;
 }
+
+// Stacking DAO is the one pool that cannot be staked against pox-5 directly:
+// its signer-manager's validate-stake! only passes while their native-pool
+// wrapper holds a transient is-delegating flag, so stake/stake-update/unstake
+// must be sent to the wrapper as delegate/delegate-update/undelegate (same
+// trailing args; delegate drops start-burn-ht and signer-calldata, which the
+// wrapper supplies itself). Their claim-staker-rewards is also non-standard:
+// it takes no staker principal and always claims for tx-sender. Everything
+// about this bespoke flow is keyed off the map below; no other pool should
+// ever be added here.
+const stackingDaoWrapperBySignerManager: Record<string, string> = {
+  'SP4SZE494VC2YC5JYG7AYFQ44F5Q4PYV7DVMDPBG.native-pool-signer-manager':
+    'SP4SZE494VC2YC5JYG7AYFQ44F5Q4PYV7DVMDPBG.native-pool-v1',
+};
+
+export function getStackingDaoWrapperContract(signerManagerContractId: string): string | undefined {
+  return stackingDaoWrapperBySignerManager[signerManagerContractId];
+}
+
+export function isStackingDaoSignerManager(signerManagerContractId: string): boolean {
+  return signerManagerContractId in stackingDaoWrapperBySignerManager;
+}
+
+export function isStackingDaoWrapperContract(contractId: string): boolean {
+  return Object.values(stackingDaoWrapperBySignerManager).includes(contractId);
+}
